created robot api
