+++ src/lastgame.txt
@@ -1,1 +1,1 @@
-r4rk1/1pqNbppp/p1n1pn2/8/P1BP4/2N5/1P2QPPP/R1BR2K1 b - - 0 13
+rnbqkbnr/pppppppp/8/8/3P4/8/PPP1PPPP/RNBQKBNR b KQkq - 0 1
